Update to cell-counting, sharptrack function and analysis
Particle size changed from 0-400 to 0-50.
Plots in Analyze_ROIs_new.m are closed immediately and saved to another folder.
Output of analysis is saved in both .mat and .csv files.
Added a button to the sharptrack atlas window "y" which saves the current window.
Added code that is commented out to only show the overlay lines in the atlas window without the background image.